fixed virtual scroll and moved mock data to external file
