+++ exadel-trainee-app/package.json
@@ -25,6 +25,7 @@
     "@angular/platform-browser-dynamic": "~12.1.1",
     "@angular/router": "~12.1.1",
     "material-icons": "^0.7.6",
+    "ng-cdk-table-virtual-scroll": "^1.0.2",
     "rxjs": "~6.6.0",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
