+++ exadel-trainee-app/src/app/models/admin-profile/admin-profile.component.html
@@ -1,75 +1,77 @@
+<div class="btn-container">
+  <button mat-raised-button color="primary" (click)="onClickNotAssign()">
+    {{ !notAssign ? 'Not Assign' : 'Assign' }}
+  </button>
+</div>
+
+<p></p>
 <cdk-virtual-scroll-viewport itemSize="500" style="height: 100%">
-  <mat-tab-group>
-    <mat-tab label="First">
-      <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-        <ng-container matColumnDef="Position">
-          <th mat-header-cell *matHeaderCellDef>No.</th>
-          <td mat-cell *matCellDef="let element">{{ element.position }}</td>
-        </ng-container>
+  <div *ngIf="notAssign">
+    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
+      <ng-container matColumnDef="Position">
+        <th mat-header-cell *matHeaderCellDef>No.</th>
+        <td mat-cell *matCellDef="let element">{{ element.position }}</td>
+      </ng-container>
 
-        <ng-container matColumnDef="Level">
-          <th mat-header-cell *matHeaderCellDef>Name</th>
-          <td mat-cell *matCellDef="let element">{{ element.level }}</td>
-        </ng-container>
+      <ng-container matColumnDef="Level">
+        <th mat-header-cell *matHeaderCellDef>Name</th>
+        <td mat-cell *matCellDef="let element">{{ element.level }}</td>
+      </ng-container>
 
-        <ng-container matColumnDef="Date">
-          <th mat-header-cell *matHeaderCellDef>Weight</th>
-          <td mat-cell *matCellDef="let element">{{ element.date | date }}</td>
-        </ng-container>
+      <ng-container matColumnDef="Date">
+        <th mat-header-cell *matHeaderCellDef>Weight</th>
+        <td mat-cell *matCellDef="let element">{{ element.date | date }}</td>
+      </ng-container>
 
-        <ng-container matColumnDef="button">
-          <th mat-header-cell *matHeaderCellDef>Symbol</th>
-          <td mat-cell *matCellDef="let element">
-            <button mat-raised-button color="primary" (click)="openDialog()">Check</button>
-          </td>
-        </ng-container>
+      <ng-container matColumnDef="button">
+        <th mat-header-cell *matHeaderCellDef>Symbol</th>
+        <td mat-cell *matCellDef="let element">
+          <button mat-raised-button color="primary" (click)="openDialog()">Check</button>
+        </td>
+      </ng-container>
 
-        <tr mat-header-row *matHeaderRowDef="displayedColumns1"></tr>
-        <tr mat-row *matRowDef="let row; columns: displayedColumns1"></tr>
-      </table>
-    </mat-tab>
-    <mat-tab label="Second">
-      <table mat-table [dataSource]="dataSource" multiTemplateDataRows class="mat-elevation-z10">
-        <ng-container matColumnDef="{{ column }}" *ngFor="let column of displayedColumns">
-          <th mat-header-cell *matHeaderCellDef>{{ column }}</th>
-          <td mat-cell *matCellDef="let element">
-            {{
-              column.toLowerCase() == 'date'
-                ? (element[column.toLowerCase()] | date)
-                : element[column.toLowerCase()]
-            }}
-          </td>
-        </ng-container>
+      <tr mat-header-row *matHeaderRowDef="displayedColumns1; sticky: true"></tr>
+      <tr mat-row *matRowDef="let row; columns: displayedColumns1"></tr>
+    </table>
+  </div>
 
-        <ng-container matColumnDef="expandedDetail">
-          <td *matCellDef="let element" [attr.colspan]="displayedColumns.length">
-            <div
-              class="example-element-detail"
-              [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'"
-            >
-              <div class="example-element-description"></div>
-              <div class="test">
-                <p>Grammar section score: {{ element.expl }}</p>
-                <button mat-raised-button color="primary">Check</button>
-              </div>
+  <div *ngIf="!notAssign">
+    <table mat-table [dataSource]="dataSource" multiTemplateDataRows class="mat-elevation-z10">
+      <ng-container matColumnDef="{{ column }}" *ngFor="let column of displayedColumns">
+        <th mat-header-cell *matHeaderCellDef>{{ column }}</th>
+        <td mat-cell *matCellDef="let element">
+          {{
+            column.toLowerCase() == 'date'
+              ? (element[column.toLowerCase()] | date)
+              : element[column.toLowerCase()]
+          }}
+        </td>
+      </ng-container>
+
+      <ng-container matColumnDef="expandedDetail">
+        <td *matCellDef="let element" [attr.colspan]="displayedColumns.length">
+          <div
+            class="example-element-detail"
+            [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'"
+          >
+            <div class="example-element-description"></div>
+            <div class="test">
+              <p>Grammar section score: {{ element.expl }}</p>
+              <button mat-raised-button color="primary">Check</button>
             </div>
-          </td>
-        </ng-container>
+          </div>
+        </td>
+      </ng-container>
 
-        <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-        <tr
-          mat-row
-          *matRowDef="let element; columns: displayedColumns"
-          class="example-element-row"
-          [class.example-expanded-row]="expandedElement === element"
-          (click)="expandedElement = expandedElement === element ? null : element"
-        ></tr>
-        <tr
-          mat-row
-          *matRowDef="let row; columns: ['expandedDetail']"
-          class="example-detail-row"
-        ></tr>
-      </table>
-    </mat-tab>
-  </mat-tab-group>
+      <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
+      <tr
+        mat-row
+        *matRowDef="let element; columns: displayedColumns"
+        class="example-element-row"
+        [class.example-expanded-row]="expandedElement === element"
+        (click)="expandedElement = expandedElement === element ? null : element"
+      ></tr>
+      <tr mat-row *matRowDef="let row; columns: ['expandedDetail']" class="example-detail-row"></tr>
+    </table>
+  </div>
 </cdk-virtual-scroll-viewport>
