+++ exadel-trainee-app/src/app/models/admin-profile/admin-profile.component.ts
@@ -2,759 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { AdminDialogComponent } from './admin-dialog/admin-dialog.component';
+import { ELEMENT_DATA, PeriodicElement } from '../../../mocks/mock_adminProfile';
 
-export interface PeriodicElement {
-  position: number;
-  date: number;
-  level: string;
-  coach: string;
-  expl: string;
-}
-
-const ELEMENT_DATA: PeriodicElement[] = [
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 1,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 2,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 3,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 4,
-    level: 'ber=g',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 5,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 6,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 7,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 8,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 9,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 10,
-    level: 'adv',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 11,
-    level: 'inter',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-  {
-    position: 12,
-    level: 'beg',
-    date: Date.now(),
-    coach: 'Nich',
-    expl: '30',
-  },
-];
 @Component({
   selector: 'app-admin-profile',
   templateUrl: './admin-profile.component.html',
@@ -768,6 +17,8 @@ const ELEMENT_DATA: PeriodicElement[] = [
   ],
 })
 export class AdminProfileComponent implements OnInit {
+  notAssign = false;
+
   constructor(public dialog: MatDialog) {}
 
   displayedColumns: string[] = ['Position', 'Level', 'Date', 'Coach'];
@@ -785,5 +36,9 @@ export class AdminProfileComponent implements OnInit {
     dialogRef.afterClosed().subscribe((result) => {
       this.dataSource[0].coach = result;
     });
+  }
+
+  onClickNotAssign(): boolean {
+    return (this.notAssign = !this.notAssign);
   }
 }
