+++ exadel-trainee-app/src/mocks/mock_adminProfile.ts
@@ -1,0 +1,752 @@
+export interface PeriodicElement {
+  position: number;
+  date: number;
+  level: string;
+  coach: string;
+  expl: string;
+}
+
+export const ELEMENT_DATA: PeriodicElement[] = [
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 1,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 2,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 3,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 4,
+    level: 'ber=g',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 5,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 6,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 7,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 8,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 9,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 10,
+    level: 'adv',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 11,
+    level: 'inter',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+  {
+    position: 12,
+    level: 'beg',
+    date: Date.now(),
+    coach: 'Nich',
+    expl: '30',
+  },
+];
